Refactor the code to work better with JSDoc.
I'm not convinced this is even advisable. Just want to get some feedback
on the way the code is now structured. It still adheres to a more
functional style, with no use of constructors, `this` or `new`. But what
we end up with is a semi pointless Client object whose primary purpose
is to communicate the Client interface to users. Its usage in code is
limited to acting as the Prototype for an actual client object
instance. Is this useful or even advisable?