Refactor the API to be purely functional.
All of this struggling with JSDoc and the solutions I had for
documenting the code, made everything look like it was trying not to be
OOP, but failing in the attempt. Realizing that even having a creator
function that returns some kind of client is unnecessary. The object
returned from this only had two functions hanging off of it. These
functions don't really need any private data outside of what can be
provided by the user.

Previously, a user had to create an options object to provide to the
`apiClient` function. From this object, I extracted an endpoint and an
access token and then used that data within a closure around the
functions that need it (e.g. `doGet()`). But since the user is creating
an object almost identical to the client, just without the functions,
why not just pass this user-created options object into each API
function.

All of this makes the code much simpler - I deleted a lot of code! And
is also more truly functional. Nothing really has state this way.